- Compile fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@31988 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/watermesh/object/watermesh.cpp
@@ -953,7 +953,7 @@ csRef<iTextureWrapper> csWaterMeshObjectFactory::MakeFresnelTex(int size)
       }
       else
       {
-        ratio = sqrt(dist / (maxDist * maxDist));
+        ratio = sqrt((float)dist / (maxDist * maxDist));
         g = sqrt((ratio * ratio) - 1.0 + (n * n));
         a = g - ratio;
         b = g + ratio;
